CODE CORRECTION: defensive event gathering
Corrects a bug when aggregating the defensive events for players, which hadn't grouped by their position in the match. It meant that, downstream, a player who had played as a Winger and Wide Midfielder would have an entry per position, with correct minutes played for each position, but with the same (total) defensive actions values.

Luckily, the correction doesn't affect the main direction of the results, but some of the more peculiar trends (like Wing-Backs being a source of interest) are no longer present